# docs/openapi/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SolarMetrics API",
    "version": "v1",
    "description": "API REST para monitoramento de energia solar. Sprint 4: paginação, filtros, HATEOAS, JWT, MongoDB (chatbot)."
  },
  "paths": {
...
    "/Cliente": {
      "get": {
        "tags": ["Cliente"],
        "summary": "Listar clientes (paginado, HATEOAS)",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortDir", "in": "query", "schema": { "type": "string" } },
          { "name": "nome", "in": "query", "schema": { "type": "string" } },
          { "name": "email", "in": "query", "schema": { "type": "string" } },
          { "name": "tipoUsuario", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "PagedResource com _links" } }
      },
      "post": { "tags": ["Cliente"], "summary": "Criar cliente", "responses": { "201": { "description": "Created" } } },
      "put": { "tags": ["Cliente"], "summary": "Atualizar cliente", "responses": { "200": { "description": "OK" } } }
    },
    "/Cliente/{id}": {
      "get": { "tags": ["Cliente"], "summary": "Buscar por id (HATEOAS)", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "ApiResource" } } },
      "delete": { "tags": ["Cliente"], "summary": "Remover", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "204": { "description": "No Content" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "Bearer": [] }]
}